pyproject.toml
[project]
name = "wnix"
version = "0.0.1"
authors = [
  { name="Jason Wilkes", email="notarealdeveloper@gmail.com" },
]
description = "Unix for software 2.0"
dependencies = [
    "requests",
    "bs4",
    "fleep",
    "validators",
    "numpy",
    "Pillow",
    "pypdf",
    "sentence_transformers",
    "transformers",
    "pytest",
    "assure",
    "rootfs",
]

[project.urls]
"Homepage" = "https://github.com/notarealdeveloper/wnix"

[project.scripts]
input       = "wnix.bin.input:main"
input-pdf   = "wnix.bin.input.pdf:main"
input-image = "wnix.bin.input.image:main"
input-url   = "wnix.bin.input.url:main"
input-html  = "wnix.bin.input.html:main"
pdf         = "wnix.bin.input.pdf:main"
image       = "wnix.bin.input.image:main"
url         = "wnix.bin.input.url:main"
html        = "wnix.bin.input.html:main"
think       = "wnix.bin.middle.think:main"
output      = "wnix.bin.output:main"
what        = "wnix.bin.output.what:main"
